requirements.txt
aiogram==2.23.1
pyyaml==6.0.1
asyncio==3.4.3
websockets==12.0
python-dotenv==1.0.0